--- a/Fletcher32.cpp
+++ b/Fletcher32.cpp
@@ -29,9 +29,17 @@ void Fletcher32::add(uint16_t value)
 {
   _count++;
   _s1 += value;
-   if (_s1 >= FLETCHER_32) _s1 -= FLETCHER_32;
+#ifdef ARDUINO_ARCH_SAMD
+  _s1 = (_s1 & 65535UL) + (_s1 >> 16);
+#else
+  if (_s1 >= FLETCHER_32) _s1 -= FLETCHER_32;
+#endif
   _s2 += _s1;
-   if (_s2 >= FLETCHER_32) _s2 -= FLETCHER_32;
+#ifdef ARDUINO_ARCH_SAMD
+  _s2 = (_s2 & 65535UL) + (_s2 >> 16);
+#else
+  if (_s2 >= FLETCHER_32) _s2 -= FLETCHER_32;
+#endif
 }
 
 //  NOTE: padding with zero's error