@daniel-mohr 2022-09-07
Sorry, this PR is based on #6. I hope #6 is OK and therefore also this dependency here.