@RobTillaart 2022-09-07
As it is rather late I will dive into the proposed code tomorrow.
The numbers are not always an improvement, so it might need conditional code to select fastest method per platform.
If AVR ..erc