@RobTillaart 2022-09-08
Looks good, merging this makes PR #6 obsolete.